# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CyberCultX Human Risk Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List report jobs
      parameters:
        - in: query
          name: type
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReportJob"
  /reports/generate:
    post:
      operationId: generateReport
      tags: [reports]
      summary: Generate a report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportGenerateInput"
      responses:
        "200":
          description: Report job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportJob"
  /reports/{id}/download:
    get:
      operationId: downloadReport
      tags: [reports]
      summary: Download a report file
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report file download (mock JSON)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportJob"
        "404":
          description: Not found
...
  /predictive/org:
    get:
      operationId: getPredictiveOrg
      tags: [predictive]
      summary: Get org-level predictive risk indicators
      responses:
        "200":
          description: Org predictive risk indicators
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PredictiveOrgResponse"
        "403": { description: Forbidden }
  /predictive/employee/{id}:
    get:
      operationId: getPredictiveEmployee
      tags: [predictive]
      summary: Get employee-level predictive risk indicators
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Employee predictive risk indicators
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PredictiveEmployeeResponse"
        "403": { description: Forbidden }
        "404": { description: Employee not found }
  /reports/arabic-pdf:
    get:
      operationId: getArabicPdfReport
      tags: [reports]
      summary: Generate Arabic RTL PDF security intelligence report
      responses:
        "200":
          description: PDF report (RTL Arabic)
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "403": { description: Forbidden }
components:
  schemas:
...
    ReportJob:
      type: object
      required: [id, type, status, format, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        status:
          type: string
        format:
          type: string
        filters:
          type: ["object", "null"]
        fileUrl:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
    ReportGenerateInput:
      type: object
      required: [type]
      properties:
        type:
          type: string
        format:
          type: string
        filters:
          type: object
...
    PredictiveIndicator:
      type: object
      required: [key, score, level, confidence, trend]
      properties:
        key:
          type: string
        score:
          type: number
        level:
          type: string
        confidence:
          type: number
        trend:
          type: string
        description:
          type: string
    PredictiveOrgResponse:
      type: object
      required: [indicators, meta]
      properties:
        indicators:
          type: array
          items:
            $ref: "#/components/schemas/PredictiveIndicator"
        meta:
          type: object
          properties:
            totalUsers:
              type: integer
            usersWithProfile:
              type: integer
            usersWithCci:
              type: integer
            computedAt:
              type: string
    PredictiveEmployeeResponse:
      type: object
      required: [userId, indicators, meta]
      properties:
        userId:
          type: integer
        indicators:
          type: array
          items:
            $ref: "#/components/schemas/PredictiveIndicator"
        meta:
          type: object
          properties:
            hasProfile:
              type: boolean
            hasCci:
              type: boolean
            hasPhish:
              type: boolean
            totalPhishEvents:
              type: integer
            computedAt:
              type: string